src/repo_crawler/requirements.txt
gitpython
azure-storage-blob
python-dotenv